Various cleanup, fixes, improvements:
- Will prompt for login name / password so they do not need to be saved
into the script.  Password will be prompted without echoing if possible.
This can be left blank if there is a stored token.

- Will prompt for search term so this is not hardcoded as めぐみん.

- Fix the more obvious issues that would arise were multiple instances of
the class to be created, because class attributes are added to instances
by reference, not copied (not tested as of yet).

- The fatal() method called the exit() function which is only supposed to
be used in interactive mode. Use sys.exit() instead, which is the one
supposed to be used by scripts. Also, exit with status 1 (error condition).

- Store login name with the access token so a different login name can be
supplied later without manually deleting the atoken file. Will still use
old-format atoken file if present provided that login name and password
are not supplied.

- Add a .gitignore file so the atoken file can't be accidentally committed
to Git (because that would give everyone access to your account).

- Fix extraneous whitespace introduced into trial_url in test_access_token
in an apparent attempt to wrap a long line, by making it a concatenation
of two literals (which can accordingly be wrapped properly).

- Wrap or otherwise split other particularly long lines.